+++ packages/react/src/components/AccountSettings/ConfigureMFA/defaults/EnableMFAButton.tsx
@@ -1,0 +1,3 @@
+import { Button } from '../../../../primitives/';
+
+export const EnableMFAButton = Button;
