+++ packages/react/src/components/AccountSettings/ConfigureMFA/defaults/index.ts
@@ -1,0 +1,6 @@
+export * from './ConfigureSMS';
+export * from './ConfigureTOTP';
+export * from './DisplayCurrentMFA';
+export * from './EnableMFAButton';
+export * from './SelectMFA';
+export * from './VerifySMS';
